requirements.txt
requests
BeautifulSoup4
python-telegram-bot
